Fix: remove duplicated codes
