+++ tests/Neo.Compiler.CSharp.UnitTests/SecurityAnalyzer/UnitTest_MissingCheckWitness.cs
@@ -63,7 +63,7 @@ public class Contract : SmartContract
     }
 }";
 
-            var context = CompileSingleContract(source);
+            var context = TestHelper.CompileSingleContract(source);
             Assert.IsTrue(context.Success, string.Join(Environment.NewLine, context.Diagnostics.Select(p => p.ToString())));
 
             var result = MissingCheckWitnessAnalyzer.AnalyzeMissingCheckWitness(
@@ -93,7 +93,7 @@ public class Contract : SmartContract
     }
 }";
 
-            var context = CompileSingleContract(source);
+            var context = TestHelper.CompileSingleContract(source);
             Assert.IsTrue(context.Success, string.Join(Environment.NewLine, context.Diagnostics.Select(p => p.ToString())));
 
             var result = MissingCheckWitnessAnalyzer.AnalyzeMissingCheckWitness(
@@ -103,39 +103,6 @@ public class Contract : SmartContract
 
             Assert.IsFalse(result.vulnerableMethodNames.Contains("_deploy"));
             Assert.IsFalse(result.vulnerableMethodNames.Contains("_initialize"));
-        }
-
-        private static CompilationContext CompileSingleContract(string sourceCode)
-        {
-            var tempFile = Path.Combine(Path.GetTempPath(), $"{Guid.NewGuid():N}.cs");
-            File.WriteAllText(tempFile, sourceCode);
-
-            try
-            {
-                var options = new CompilationOptions
-                {
-                    Optimize = CompilationOptions.OptimizationType.All,
-                    Nullable = Microsoft.CodeAnalysis.NullableContextOptions.Enable,
-                    SkipRestoreIfAssetsPresent = true
-                };
-
-                var engine = new CompilationEngine(options);
-                var repoRoot = SyntaxProbeLoader.GetRepositoryRoot();
-                var frameworkProject = Path.Combine(repoRoot, "src", "Neo.SmartContract.Framework", "Neo.SmartContract.Framework.csproj");
-
-                var contexts = engine.CompileSources(new CompilationSourceReferences
-                {
-                    Projects = new[] { frameworkProject }
-                }, tempFile);
-
-                Assert.AreEqual(1, contexts.Count, "Expected exactly one contract compilation context.");
-                return contexts[0];
-            }
-            finally
-            {
-                if (File.Exists(tempFile))
-                    File.Delete(tempFile);
-            }
         }
     }
 }
