+++ tests/Neo.Compiler.CSharp.UnitTests/UnitTest_DirectInitDiagnostics.cs
@@ -45,7 +45,7 @@ public class UnitTest_DirectInitDiagnostics
 
     private static CompilationContext CompileSingleContractWithUInt256Literal(string literal)
     {
-        return CompileSingleContract($$"""
+        return TestHelper.CompileSingleContract($$"""
 using Neo.SmartContract.Framework;
 
 public class Contract : SmartContract
@@ -58,38 +58,5 @@ public class Contract : SmartContract
     }
 }
 """);
-    }
-
-    private static CompilationContext CompileSingleContract(string sourceCode)
-    {
-        var tempFile = Path.Combine(Path.GetTempPath(), $"{Guid.NewGuid():N}.cs");
-        File.WriteAllText(tempFile, sourceCode);
-
-        try
-        {
-            var options = new CompilationOptions
-            {
-                Optimize = CompilationOptions.OptimizationType.All,
-                Nullable = NullableContextOptions.Enable,
-                SkipRestoreIfAssetsPresent = true
-            };
-
-            var engine = new CompilationEngine(options);
-            var repoRoot = SyntaxProbeLoader.GetRepositoryRoot();
-            var frameworkProject = Path.Combine(repoRoot, "src", "Neo.SmartContract.Framework", "Neo.SmartContract.Framework.csproj");
-
-            var contexts = engine.CompileSources(new CompilationSourceReferences
-            {
-                Projects = new[] { frameworkProject }
-            }, tempFile);
-
-            Assert.AreEqual(1, contexts.Count, "Expected exactly one contract compilation context.");
-            return contexts[0];
-        }
-        finally
-        {
-            if (File.Exists(tempFile))
-                File.Delete(tempFile);
-        }
     }
 }
