+++ tests/Neo.Compiler.CSharp.UnitTests/UnitTest_SwitchJumpTable.cs
@@ -110,39 +110,6 @@ public class Contract : SmartContract
         AssertUsesOptimizedDispatch(source, "Contract.Main(int)");
     }
 
-    private static CompilationContext CompileSingleContract(string sourceCode)
-    {
-        var tempFile = Path.Combine(Path.GetTempPath(), $"{Guid.NewGuid():N}.cs");
-        File.WriteAllText(tempFile, sourceCode);
-
-        try
-        {
-            var options = new CompilationOptions
-            {
-                Optimize = CompilationOptions.OptimizationType.Basic,
-                Nullable = NullableContextOptions.Enable,
-                SkipRestoreIfAssetsPresent = true
-            };
-
-            var engine = new CompilationEngine(options);
-            var repoRoot = Syntax.SyntaxProbeLoader.GetRepositoryRoot();
-            var frameworkProject = Path.Combine(repoRoot, "src", "Neo.SmartContract.Framework", "Neo.SmartContract.Framework.csproj");
-
-            var contexts = engine.CompileSources(new CompilationSourceReferences
-            {
-                Projects = new[] { frameworkProject }
-            }, tempFile);
-
-            Assert.AreEqual(1, contexts.Count, "Expected exactly one contract compilation context.");
-            return contexts[0];
-        }
-        finally
-        {
-            if (File.Exists(tempFile))
-                File.Delete(tempFile);
-        }
-    }
-
     private static string BuildSwitchContract(string inputType, string[] caseLabels)
     {
         var cases = string.Join(Environment.NewLine, caseLabels.Select((label, index) => $"            case {label}: return {index};"));
@@ -181,7 +148,7 @@ public class Contract : SmartContract
 
     private static OpCode[] GetMethodOpcodes(string source, string methodId)
     {
-        var context = CompileSingleContract(source);
+        var context = TestHelper.CompileSingleContract(source);
         Assert.IsTrue(context.Success, string.Join(Environment.NewLine, context.Diagnostics.Select(p => p.ToString())));
 
         var nef = context.CreateExecutable();
