--- tests/Neo.Compiler.CSharp.UnitTests/Utils/Extensions.cs
@@ -1,20 +1,0 @@
-// Copyright (C) 2015-2026 The Neo Project.
-//
-// Extensions.cs file belongs to the neo project and is free
-// software distributed under the MIT software license, see the
-// accompanying file LICENSE in the main directory of the
-// repository or http://www.opensource.org/licenses/mit-license.php
-// for more details.
-//
-// Redistribution and use in source and binary forms with or without
-// modifications are permitted.
-
-using System.IO;
-
-namespace Neo.Compiler.CSharp.UnitTests.Utils
-{
-    static class Extensions
-    {
-        public static readonly string TestContractRoot = Path.GetFullPath("../../../../Neo.Compiler.CSharp.TestContracts/") + "/";
-    }
-}
